% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_summarize_schema}
\alias{.check_summarize_schema}
\title{Schema canary for summarize() output}
\usage{
.check_summarize_schema(result)
}
\description{
Schema canary for summarize() output
}
\keyword{internal}
